requirements.txt
gspread
google-auth
openai
requests
google-api-python-client
google-auth-oauthlib
google-auth-httplib2